feat(project-setup): cLI that displays formatted JSON data in a styled, localized table
This project is a Command-Line Interface (CLI) that reads data from a JSON file, formats the
information based on the locale (pt-BR), and displays it in a styled table using libraries like
chalk, chalk-table, and draftlog. It also enables basic user interaction through readline, capturing
input and logging responses in the terminal. The initial setup focuses on clean structure, data
formatting, and internationalization (i18n) support.

#21